requirements.txt
transformers
gradio
pyttsx3
Pillow
torch
gtts
sentencepiece
accelerate
